requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
comm==0.1.2
crispy-bootstrap4==2022.1
debugpy==1.6.6
decorator==5.1.1
Django==4.1.7
django-crispy-forms==2.0
executing==1.2.0
ipykernel==6.21.3
ipython==8.11.0
jedi==0.18.2
jupyter_client==8.0.3
jupyter_core==5.2.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
packaging==23.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
pywin32==305
pyzmq==25.0.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
tzdata==2022.7
wcwidth==0.2.6
